requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.5.1
apache-airflow-providers-common-io==1.5.1
apache-airflow-providers-common-sql==1.24.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.1
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.1
apache-airflow-providers-sqlite==4.0.1
apispec==6.8.1
argcomplete==3.6.1
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
bert-score==0.3.13
blinker==1.9.0
blis==1.2.1
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
cloudpathlib==0.21.0
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-re2==1.1.20240702
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.23
langchain-anthropic==0.3.12
langchain-core==0.3.54
langchain-huggingface==0.1.2
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.19
language_data==1.3.0
lazy-object-proxy==1.10.0
limits==4.4.1
linkify-it-py==2.0.3
load-dotenv==0.1.0
lockfile==0.12.2
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
murmurhash==1.0.12
narwhals==1.35.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
ordered-set==4.1.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.17.0
pendulum==3.0.0
pillow==11.1.0
plotly==6.0.1
pluggy==1.5.0
poppler-utils==0.1.0
portalocker==2.10.1
preshed==3.0.9
prison==0.2.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.4
pymupdf4llm==0.0.19
pyparsing==3.2.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.13.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
srsly==2.5.1
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
time-machine==2.16.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==2.2.3
wirerope==1.0.0
wordcloud==1.9.4
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
